ver4.1.7 commit

+++ cszp/cszp_soccer.py
@@ -10,7 +10,7 @@ from texttable import *
 from cszp import cszp_html, cszp_log
 
 
-def loop(cmd, loops, lang):
+def loop(cmd, loops, lang, module):
     subp.reset()
     inp = ""
     while inp.lower() != "y" and inp.lower() != "n":
@@ -83,16 +83,7 @@ def loop(cmd, loops, lang):
             pass
         logs += "SERVER-CMD : " + sc + "\n"
 
-        try:
-            data = open("./config/config.conf", "r")
-        except subprocess.CalledProcessError:
-            data = open("./config/config.conf", "w")
-            data.write(
-                "soccerwindow2start,on,automake,off,rcglog output,on,rcllog output,on,logfile output," + os.path.expanduser(
-                    "~") +
-                "/csvdata")
-            data.close()
-            data = open("./config/config.conf", "r")
+        data = module.Open("./config/config.conf")
         datas = data.read()
         data.close()
         datas = datas.split(",")
@@ -140,17 +131,7 @@ def loop(cmd, loops, lang):
         except subprocess.CalledProcessError:
             pass
         if i1 == "y":
-            try:
-                data = open("./config/config.conf", "r")
-            except FileNotFoundError:
-                data = open("./config/config.conf", "w")
-                data.write(
-                    "soccerwindow2start,on,automake,off,rcglog output,on,rcllog output,on,logfile output," + os.path.expanduser(
-                        "~")
-                    + "/csvdata"
-                )
-                data.close()
-                data = open("./config/config.conf", "r")
+            data = module.Open("./config/config.conf")
             df = data.read()
             # print(df.split("'"))
             temp = open(df.split(",")[9] + "/" + exittime + "_team1.log", "w")
@@ -160,34 +141,14 @@ def loop(cmd, loops, lang):
             temp.write(logt2)
             temp.close()
         if i2 == "y":
-            try:
-                data = open("./config/config.conf", "r")
-            except FileNotFoundError:
-                data = open("./config/config.conf", "w")
-                data.write(
-                    "soccerwindow2start,on,automake,off,rcglog output,on,rcllog output,on,logfile output," + os.path.expanduser(
-                        "~")
-                    + "/csvdata"
-                )
-                data.close()
-                data = open("./config/config.conf", "r")
+            data = module.Open("./config/config.conf")
             df = data.read()
             # print(df)
             temp = open(df.split(",")[9] + "/" + exittime + "_server.log", "w")
             temp.write(logs)
             temp.close()
         if i3 == "y":
-            try:
-                data = open("./config/config.conf", "r")
-            except FileNotFoundError:
-                data = open("./config/config.conf", "w")
-                data.write(
-                    "soccerwindow2start,on,automake,off,rcglog output,on,rcllog output,on,logfile output," + os.path.expanduser(
-                        "~")
-                    + "/csvdata"
-                )
-                data.close()
-                data = open("./config/config.conf", "r")
+            data = module.Open("./config/config.conf")
             df = data.read()
             # print(df.split(",")[9] + "/" + file)
             temp = open(df.split(",")[9] + "/" + file, "a")
@@ -196,9 +157,12 @@ def loop(cmd, loops, lang):
                 "\n" + exittime + "," + soccer[0] + "," + soccer[1] + "," + soccer[2] + "," + soccer[3] + "," + str(
                     int(soccer[2]) - int(soccer[3])) + "," + str(int(soccer[3]) - int(soccer[2])))
             temp.close()
-            cszp_html.plot_socre(file)
+            cszp_html.plot_socre(file, module)
 
-            temp = open("./csvdata/data.csv", "a")
+            data = module.Open("./config/config.conf")
+            path = data.read().split(",")[9]
+            data.close()
+            temp = open(path + "/data.csv", "a")
             soccer = cszp_log.log(logs)
             temp.write(
                 "\n" + exittime + "," + soccer[0] + "," + soccer[1] + "," + soccer[2] + "," + soccer[3] + "," + str(
@@ -207,13 +171,14 @@ def loop(cmd, loops, lang):
 
     print("\033[1m\033[38;5;172m---------- " + datetime.now().strftime("%Y/%m/%d %H:%M:%S") + " -----------")
     # print(serverpid)
-    subprocess.check_output("kill " + serverpid, shell=True)
-    # time.sleep(2)
-    cszp_html.logs(exittime)
+    if i3 == "y":
+        subprocess.check_output("kill " + serverpid, shell=True)
+        # time.sleep(2)
+        cszp_html.logs(exittime)
     subp.Input(lang.lang("Enterキーを押して続行..."))
 
 
-def start(cmd, lang):
+def start(cmd, lang, module):
     subp.reset()
     logs = ""
     logt1 = ""
@@ -241,16 +206,7 @@ def start(cmd, lang):
     print("\033[38;5;2mSERVER-CMD : " + sc)
     print("WINDOW-CMD : " + wc + "\033[0m")
 
-    try:
-        data = open("./config/config.conf", "r")
-    except FileNotFoundError:
-        data = open("./config/config.conf", "w")
-        data.write(
-            "soccerwindow2start,on,automake,off,rcglog output,on,rcllog output,on,logfile output," + os.path.expanduser(
-                "~") +
-            "/csvdata")
-        data.close()
-        data = open("./config/config.conf", "r")
+    data = module.Open("./config/config.conf")
     datas = data.read()
     data.close()
     datas = datas.split(",")
@@ -347,23 +303,14 @@ def start(cmd, lang):
     except subprocess.CalledProcessError:
         pass
     soccer = cszp_log.log(logs)
-    print(soccer[0] + " " + soccer[2] + " - " + soccer[3] + " " + soccer[1] + " " + cszp_log.kekka(logs, 0))
+    print(soccer[0] + " " + soccer[2] + " - " + soccer[3] + " " + soccer[1] + " " + cszp_log.kekka(logs, 0, module))
     inp = ""
     while inp.lower() != "y" and inp.lower() != "n":
         inp = subp.Input(lang.lang("プログラムログを保存しますか？ Y/n"))
         if inp.lower() != "y" and inp.lower() != "n":
             print("\033[1m\033[38;5;9mERR:\033[0m" + lang.lang("yまたはnで答えてください"))
     if inp.lower() == "y":
-        try:
-            data = open("./config/config.conf", "r")
-        except FileNotFoundError:
-            data = open("./config/config.conf", "w")
-            data.write(
-                "soccerwindow2start,on,automake,off,rcglog output,on,rcllog output,on,logfile output," + os.path.expanduser(
-                    "~") +
-                "/csvdata")
-            data.close()
-            data = open("./config/config.conf", "r")
+        data = module.Open("./config/config.conf")
         df = data.read()
         temp = open(df.split(",")[9] + "/" + exittime + "_team1.log", "w")
         temp.write(logt1)
@@ -377,16 +324,7 @@ def start(cmd, lang):
         if inp.lower() != "y" and inp.lower() != "n":
             print("\033[1m\033[38;5;9mERR:\033[0m" + lang.lang("yまたはnで答えてください"))
     if inp.lower() == "y":
-        try:
-            data = open("./config/config.conf", "r")
-        except FileNotFoundError:
-            data = open("./config/config.conf", "w")
-            data.write(
-                "soccerwindow2start,on,automake,off,rcglog output,on,rcllog output,on,logfile output," + os.path.expanduser(
-                    "~") +
-                "/csvdata")
-            data.close()
-            data = open("./config/config.conf", "r")
+        data = module.Open("./config/config.conf")
         df = data.read()
         temp = open(df.split(",")[9] + "/" + exittime + "_server.log", "w")
         temp.write(logs)
@@ -399,16 +337,7 @@ def start(cmd, lang):
             print("\033[1m\033[38;5;9mERR:\033[0m" + lang.lang("yまたはnで答えてください"))
     if inp.lower() == "y":
         inp = subp.Input("FileName?")
-        try:
-            data = open("./config/config.conf", "r")
-        except FileNotFoundError:
-            data = open("./config/config.conf", "w")
-            data.write(
-                "soccerwindow2start,on,automake,off,rcglog output,on,rcllog output,on,logfile output," + os.path.expanduser(
-                    "~") +
-                "/csvdata")
-            data.close()
-            data = open("./config/config.conf", "r")
+        data = module.Open("./config/config.conf")
         df = data.read()
         data.close()
         temp = open(df.split(",")[9] + "/" + inp, "a")
@@ -418,7 +347,10 @@ def start(cmd, lang):
                 int(soccer[2]) - int(soccer[3])) + "," + str(int(soccer[3]) - int(soccer[2])))
         temp.close()
 
-    temp = open("./csvdata/data.csv", "a")
+    data = module.Open("./config/config.conf")
+    path = data.read().split(",")[9]
+    data.close()
+    temp = open(path + "/data.csv", "a")
     soccer = cszp_log.log(logs)
     temp.write(
         "\n" + exittime + "," + soccer[0] + "," + soccer[1] + "," + soccer[2] + "," + soccer[3] + "," + str(
@@ -432,17 +364,7 @@ def start(cmd, lang):
             if inp.lower() != "y" and inp.lower() != "n":
                 print("\033[1m\033[38;5;9mERR:\033[0m" + lang.lang("yまたはnで答えてください"))
         if inp.lower() == "y":
-            try:
-                data = open("./config/config.conf", "r")
-            except FileNotFoundError:
-                data = open("./config/config.conf", "w")
-                data.write(
-                    "soccerwindow2start,on,automake,off,rcglog output,on,rcllog output,on,logfile output," + os.path.expanduser(
-                        "~")
-                    + "/csvdata"
-                )
-                data.close()
-                data = open("./config/config.conf", "r")
+            data = module.Open("./config/config.conf")
             df = data.read()
             data.close()
             temp = open(df.split(",")[9] + "/" + exittime + "_makelog.log", "w")
@@ -450,7 +372,7 @@ def start(cmd, lang):
             temp.close()
 
 
-def setting(lang, testmode=False, loopmode=False):
+def setting(lang, module, testmode=False, loopmode=False):
     ok = 0
     cmd = []
 
@@ -462,16 +384,7 @@ def setting(lang, testmode=False, loopmode=False):
     print("\n\n\033[1m\033[38;5;10m" + lang.lang("cszp 簡単サッカー実行プログラム"))
     print("\033[38;5;39m" + lang.lang("設定確認"))
     # noinspection PyBroadException
-    try:
-        data = open("./config/config.conf", "r")
-    except Exception:
-        data = open("./config/config.conf", "w")
-        data.write(
-            "soccerwindow2start,on,automake,off,rcglog output,on,rcllog output,on,logfile output," + os.path.expanduser(
-                "~") +
-            "/csvdata")
-        data.close()
-        data = open("./config/config.conf", "r")
+    data = module.Open("./config/config.conf")
     datas = data.read()
     data.close()
     datas = datas.split(",")
@@ -488,20 +401,9 @@ def setting(lang, testmode=False, loopmode=False):
     subp.reset()
     while ok == 0:
         subprocess.check_call("clear", shell=True)
-        print("\033[0m\033[38;5;172m")
-        v = open("./version")
-        subprocess.check_call("figlet -ctk cszp " + v.read(), shell=True)
-        v.close()
-        print("\n\n\033[1m\033[38;5;10mcszp " + lang.lang("cszp 簡単サッカー実行プログラム"))
-        print("\n\033[38;5;39m" + lang.lang("簡単サッカー実行リスト"))
+
         # noinspection PyBroadException
-        try:
-            data = open("./config/setting.conf", "r")
-        except Exception:
-            data = open("./config/setting.conf", "w")
-            data.write("name,command")
-            data.close()
-            data = open("./config/setting.conf", "r")
+        data = module.Open("./config/setting.conf")
         datas = data.read()
         data.close()
         datas = datas.split(",")
@@ -512,8 +414,20 @@ def setting(lang, testmode=False, loopmode=False):
             datal.append(datat)
         table = Texttable()
         table.add_rows(datal)
-        print(table.draw() + lang.lang("\n前画面に戻る場合はbackと入力\n"))
-        inp = subp.Inputfile(lang.lang("\nTeam1(黄色)チームのパスまたはリストの名前を入力"), textcolor="\033[38;5;11m")
+        # inp = subp.Inputfile(lang.lang("\nTeam1(黄色)チームのパスまたはリストの名前を入力"), textcolor="\033[38;5;11m")
+        print("\033[0m")
+        inp = subp.Inputfilegui(lang.lang("Team1(黄色)チームのパスを入力(リストの名前を入力したい場合はQキー)"))
+        if inp == -1:
+            print("\033[0m\033[38;5;172m")
+            v = open("./version")
+            subprocess.check_call("figlet -ctk cszp " + v.read(), shell=True)
+            v.close()
+            print("\n\n\033[1m\033[38;5;10mcszp " + lang.lang("cszp 簡単サッカー実行プログラム"))
+            print("\n\033[38;5;39m" + lang.lang("簡単サッカー実行リスト"))
+            print(table.draw() + lang.lang("\n前画面に戻る場合はbackと入力\n"))
+            inp = subp.Input(lang.lang("\nTeam1(黄色)チームのリストの名前を入力"), textcolor="\033[38;5;11m")
+        if inp == -2:
+            raise KeyboardInterrupt
 
         if inp != "back":
             i = 0
@@ -544,20 +458,9 @@ def setting(lang, testmode=False, loopmode=False):
     subp.reset()
     while ok == 0 and ok != 2:
         subprocess.check_call("clear", shell=True)
-        print("\033[0m\033[38;5;172m")
-        v = open("./version")
-        subprocess.check_call("figlet -ctk cszp " + v.read(), shell=True)
-        v.close()
-        print("\n\n\033[1m\033[38;5;10m" + lang.lang("cszp 簡単サッカー実行プログラム"))
-        print("\n\033[38;5;39m" + lang.lang("簡単サッカー実行リスト"))
+
         # noinspection PyBroadException
-        try:
-            data = open("./config/setting.conf", "r")
-        except Exception:
-            data = open("./config/setting.conf", "w")
-            data.write("name,command")
-            data.close()
-            data = open("./config/setting.conf", "r")
+        data = module.Open("./config/setting.conf")
         datas = data.read()
         data.close()
         datas = datas.split(",")
@@ -568,9 +471,22 @@ def setting(lang, testmode=False, loopmode=False):
             datal.append(datat)
         table = Texttable()
         table.add_rows(datal)
-        print(
-            table.draw() + lang.lang("\n前画面に戻る場合はbackと入力") + "\n\033[38;5;11m" + lang.lang("Team1(黄色)チームのパス:") + cmd[0])
-        inp = subp.Inputfile(lang.lang("\nTeam2(赤色)チームのパスまたはリストの名前を入力"), textcolor="\033[38;5;9m")
+        # inp = subp.Inputfile(lang.lang("\nTeam2(赤色)チームのパスまたはリストの名前を入力"), textcolor="\033[38;5;9m")
+        print("\033[0m")
+        inp = subp.Inputfilegui(lang.lang("Team2(赤色)チームのパスを入力(リストの名前を入力したい場合はQキー)"))
+        if inp == -1:
+            print("\033[0m\033[38;5;172m")
+            v = open("./version")
+            subprocess.check_call("figlet -ctk cszp " + v.read(), shell=True)
+            v.close()
+            print("\n\n\033[1m\033[38;5;10mcszp " + lang.lang("cszp 簡単サッカー実行プログラム"))
+            print("\n\033[38;5;39m" + lang.lang("簡単サッカー実行リスト"))
+            print(
+                table.draw() + lang.lang("\n前画面に戻る場合はbackと入力") + "\n\033[38;5;11m" + lang.lang("Team1(黄色)チームのパス:") +
+                cmd[0])
+            inp = subp.Input(lang.lang("\nTeam2(赤色)チームのリストの名前を入力"), textcolor="\033[38;5;9m")
+        if inp == -2:
+            raise KeyboardInterrupt
 
         if inp != "back":
             i = 0
@@ -603,16 +519,7 @@ def setting(lang, testmode=False, loopmode=False):
         print("\n\033[38;5;11m" + lang.lang("Team1(黄色)チームのパス:") + cmd[0])
         print("\033[38;5;9m" + lang.lang("Team2(赤色)チームのパス:") + cmd[1])
         inp = subp.Input(lang.lang("\nサーバーの引数を入力（ない場合は空欄）"), textcolor="\033[38;5;9m")
-        try:
-            data = open("./config/config.conf", "r")
-        except FileNotFoundError:
-            data = open("./config/config.conf", "w")
-            data.write(
-                "soccerwindow2start,on,automake,off,rcglog output,on,rcllog output,on,logfile output," + os.path.expanduser(
-                    "~") +
-                "/csvdata")
-            data.close()
-            data = open("./config/config.conf", "r")
+        data = module.Open("./config/config.conf")
         df = data.read()
         data.close()
         inp += " server::auto_mode=true server::kick_off_wait=10 server::game_over_wait=10 server::connect_wait=1500 " \
@@ -629,16 +536,7 @@ def setting(lang, testmode=False, loopmode=False):
                 except ValueError:
                     print("\033[38;5;9m" + lang.lang("ERR:数字を入力してください"))
 
-        try:
-            data = open("./config/config.conf", "r")
-        except FileNotFoundError:
-            data = open("./config/config.conf", "w")
-            data.write(
-                "soccerwindow2start,on,automake,off,rcglog output,on,rcllog output,on,logfile output," + os.path.expanduser(
-                    "~") +
-                "/csvdata")
-            data.close()
-            data = open("./config/config.conf", "r")
+        data = module.Open("./config/config.conf")
         datas = data.read()
         datas = datas.split(",")
         data.close()
@@ -650,6 +548,6 @@ def setting(lang, testmode=False, loopmode=False):
         print("\n\033[38;5;14mloading now...")
         time.sleep(0.5)
         if loopmode:
-            loop(cmd, loops, lang)
+            loop(cmd, loops, lang, module)
         else:
-            start(cmd, lang)
+            start(cmd, lang, module)
